pyproject.toml
[project]
name = "nova"
version = "0.1.0"
description = "Nova - Personal Financial Assistant DeepAgent"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
